+++ src/components/edit/block-dnd/ChordTokenDroppable.tsx
@@ -1,15 +1,8 @@
 import { RootRef } from "@material-ui/core";
 import React, { useEffect } from "react";
-import { DropTargetMonitor, useDrop } from "react-dnd";
 import { ChordBlock } from "../../../common/ChordModel/ChordBlock";
 import { IDable } from "../../../common/ChordModel/Collection";
-import {
-    DNDChord,
-    DNDChordType,
-    DropCollector,
-    DropParams,
-    TypedDropParams,
-} from "./common";
+import { useChordDrop } from "./useChordDrop";
 
 interface ChordTokenDroppableProps {
     children: React.ReactElement;
@@ -20,19 +13,7 @@ interface ChordTokenDroppableProps {
 const ChordTokenDroppable: React.FC<ChordTokenDroppableProps> = (
     props: ChordTokenDroppableProps
 ) => {
-    const [{ isOver }, dropRef] = useDrop<DNDChord, DropParams, DropCollector>({
-        accept: DNDChordType,
-        drop: (): TypedDropParams => {
-            return {
-                type: "dropped-chord-result",
-                tokenIndex: 0,
-                blockID: props.blockID,
-            };
-        },
-        collect: (monitor: DropTargetMonitor): DropCollector => ({
-            isOver: monitor.isOver({ shallow: true }),
-        }),
-    });
+    const [dropRef, isOver] = useChordDrop(props.blockID, 0);
 
     const onDragOver = props.onDragOver;
