+++ src/components/edit/block-dnd/ChordlessTokenDroppable.tsx
@@ -1,37 +1,22 @@
 import { RootRef } from "@material-ui/core";
 import React from "react";
-import { DropTargetMonitor, useDrop } from "react-dnd";
-import {
-    ClassNameable,
-    DNDChord,
-    DNDChordType,
-    DropCollector,
-    DropParams,
-    TypedDropParams,
-} from "./common";
+import { ChordBlock } from "../../../common/ChordModel/ChordBlock";
+import { IDable } from "../../../common/ChordModel/Collection";
+import { ClassNameable } from "./common";
+import { useChordDrop } from "./useChordDrop";
 
 interface ChordlessTokenDroppableProps {
     children: React.ReactElement<ClassNameable>;
     hoverableClassName: string;
     dragOverClassName: string;
-    dropParams: DropParams;
+    blockID: IDable<ChordBlock>;
+    tokenIndex: number;
 }
 
 const ChordlessTokenDroppable: React.FC<ChordlessTokenDroppableProps> = (
     props: ChordlessTokenDroppableProps
 ): JSX.Element => {
-    const [{ isOver }, dropRef] = useDrop<DNDChord, DropParams, DropCollector>({
-        accept: DNDChordType,
-        drop: (): TypedDropParams => {
-            return {
-                type: "dropped-chord-result",
-                ...props.dropParams,
-            };
-        },
-        collect: (monitor: DropTargetMonitor): DropCollector => ({
-            isOver: monitor.isOver({ shallow: true }),
-        }),
-    });
+    const [dropRef, isOver] = useChordDrop(props.blockID, props.tokenIndex);
 
     const childClassName: string = isOver
         ? props.dragOverClassName
